refact: move notifierFlow to :ext:coroutines
