+++ ext/coroutines/src/main/java/com/jeanbarrossilva/orca/ext/coroutines/notifier/Notifier.java
@@ -13,13 +13,13 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package com.jeanbarrossilva.orca.feature.postdetails.viewmodel.notifier;
+package com.jeanbarrossilva.orca.ext.coroutines.notifier;
 
-import androidx.annotation.NonNull;
 import kotlin.coroutines.Continuation;
 import kotlinx.coroutines.flow.Flow;
 import kotlinx.coroutines.flow.FlowCollector;
 import kotlinx.coroutines.flow.StateFlow;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * No-op structure that serves the sole purpose of being emitted to a {@link
@@ -35,17 +35,17 @@ import kotlinx.coroutines.flow.StateFlow;
  */
 public final class Notifier {
   /** {@link Notifier} to be emitted after the {@link #initial} one. */
-  private static final Notifier subsequent = new Notifier("Notifier.subsequent");
+  @NotNull private static final Notifier subsequent = new Notifier("Notifier.subsequent");
 
   /** Name by which this {@link Notifier} can be identified. */
-  private final String name;
+  @NotNull private final String name;
 
   /**
    * {@link Notifier} to be first emitted to a {@link MutableStateFlowExtensions#notifierFlow()}.
    */
-  static final Notifier initial = new Notifier("Notifier.initial");
+  @NotNull static final Notifier initial = new Notifier("Notifier.initial");
 
-  private Notifier(String name) {
+  private Notifier(@NotNull String name) {
     this.name = name;
   }
 
@@ -55,20 +55,21 @@ public final class Notifier {
    * @throws IllegalArgumentException If this {@link Notifier} is neither the {@link #initial} nor
    *     the {@link #subsequent} one.
    */
+  @NotNull
   Notifier next() {
-    final Notifier successor;
+    final Notifier next;
     if (this == initial) {
-      successor = subsequent;
+      next = subsequent;
     } else if (this == subsequent) {
-      successor = initial;
+      next = initial;
     } else {
       throw new IllegalArgumentException(
           "Notifier should be either the initial or the subsequent one.");
     }
-    return successor;
+    return next;
   }
 
-  @NonNull
+  @NotNull
   @Override
   public String toString() {
     return name;
