+++ feature/post-details/src/main/java/com/jeanbarrossilva/orca/feature/postdetails/viewmodel/PostDetailsViewModel.kt
@@ -22,9 +22,9 @@ import androidx.lifecycle.viewmodel.viewModelFactory
 import com.jeanbarrossilva.loadable.flow.loadable
 import com.jeanbarrossilva.loadable.list.flow.listLoadable
 import com.jeanbarrossilva.orca.core.feed.profile.post.PostProvider
+import com.jeanbarrossilva.orca.ext.coroutines.notifier.notifierFlow
+import com.jeanbarrossilva.orca.ext.coroutines.notifier.notify
 import com.jeanbarrossilva.orca.feature.postdetails.toPostDetailsFlow
-import com.jeanbarrossilva.orca.feature.postdetails.viewmodel.notifier.notifierFlow
-import com.jeanbarrossilva.orca.feature.postdetails.viewmodel.notifier.notify
 import com.jeanbarrossilva.orca.platform.autos.theme.AutosTheme
 import com.jeanbarrossilva.orca.platform.ui.component.timeline.post.toPostPreviewFlow
 import com.jeanbarrossilva.orca.platform.ui.core.await
